requirements.txt
langgraph>=0.4.0
langchain>=0.3.0
langchain-openai>=0.3.0
langchain-core>=0.3.0
fastapi>=0.115.0
uvicorn>=0.34.0
slack-sdk>=3.34.0
python-dotenv>=1.0.0

